+++ src/VSPackage.cs
@@ -28,6 +28,7 @@ namespace CloseAllTabs
             CloseOpenDocuments.Initialize(dte, options);
             CollapseFolders.Initialize(dte, options);
             SolutionExplorerFocus.Initialize(dte, options);
+            DeleteBinFolder.Initialize(dte, options);
         }
     }
 }